good stopping point for fake mic messaging
found out how SVC_VOICEDATA works after looking at hw.so in IDA Pro. There are calls to the steam voice apis, and the output is used as is in the network message payload. The same steam_api.dll that sven uses must be called for sven to decode the data.

A simple program (steam_voice) is included that prints voice data to stdout. Steam must be running and a microphone is still required, but sven doesn't need to be running anymore.

Two new python scripts stream voice data between a client (my PC) and dedicated server (twlz). The server script writes voice data to a file and a server plugin (FakeMic.as) polls it for updates.

The plugin needs better timing for sending the packets. At the moment it sends them almost all at once, which is fine when file updates come in at the same speed as voice data is generated. But, my connection to twlz is lossy, so songs will skip and cut out sometimes. Waiting a second or so before writing packets to the plugin gives the scripts time to resend lost packets, but the file writes are still inconsistently timed.