better tolerance for bad networks
the plugin sends packets at the correct times instead of in batches as they arrive. A timestamp is added to each one from the steam_voice program, which is the time the packet was generated. The plugin uses this to calculate a delay before sending the next packet to players.

Also, the client/server script can also handled one side disconnecting/restarting, and packetId overflows are handled properly.